requirements.txt
googlesearch-python
boto3